+++ typescript/packages/plugins/solana-nfts/package.json
@@ -13,6 +13,7 @@
     "types": "./dist/index.d.ts",
     "dependencies": {
         "@goat-sdk/core": "workspace:*",
+        "@goat-sdk/wallet-solana": "workspace:*",
         "@metaplex-foundation/digital-asset-standard-api": "1.0.0",
         "@metaplex-foundation/mpl-bubblegum": "4.2.1",
         "@metaplex-foundation/umi": "^0.9.2",
