+++ typescript/packages/plugins/solana-nfts/src/parameters.ts
@@ -1,0 +1,9 @@
+import { createToolParameters } from "@goat-sdk/core";
+import { z } from "zod";
+
+export class TransferNftParameters extends createToolParameters(
+    z.object({
+        recipientAddress: z.string().describe("The address to send the NFT to"),
+        assetId: z.string().describe("The asset ID of the NFT to send"),
+    }),
+) {}
